changed DebugMetric.add() to take a long parameter instead of an int

+++ graal/com.oracle.graal.debug/src/com/oracle/graal/debug/Debug.java
@@ -243,7 +243,7 @@ public class Debug {
         public void increment() {
         }
 
-        public void add(int value) {
+        public void add(long value) {
         }
     };
